src/code/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
python-dotenv
Faker
python-multipart